fix: fixed a unmounted bug
