+++ client/components/index.js
@@ -21,3 +21,4 @@ export {default as Loading} from './Loading'
 export {default as FilterTasksByLabel} from './FilterTasksByLabel'
 export {default as Footer} from './Footer'
 export {DeleteBoard} from './DeleteBoard'
+export {default as DeleteUser} from './DeleteUser'
